Move reduction to its own module, and fix Mul impls
