+++ eyelid-match-ops/src/primitives/poly/modular_poly.rs
@@ -13,42 +13,17 @@
 
 use std::ops::{Index, IndexMut, Mul};
 
-use ark_ff::{One, Zero};
-use ark_poly::{
-    polynomial::{
-        univariate::{DenseOrSparsePolynomial, DensePolynomial},
-        Polynomial,
-    },
-    univariate::SparsePolynomial,
+use ark_ff::Zero;
+use ark_poly::polynomial::univariate::{
+    DenseOrSparsePolynomial, DensePolynomial, SparsePolynomial,
 };
-use derive_more::{
-    Add, AsRef, Deref, DerefMut, Div, DivAssign, Into, Mul, MulAssign, Neg, Rem, RemAssign,
-};
-use lazy_static::lazy_static;
+use derive_more::{Add, AsRef, Deref, DerefMut, Div, Into, Neg, Rem};
 
-use crate::primitives::poly::{Coeff, MAX_POLY_DEGREE};
+use crate::primitives::poly::{mod_poly, mul_poly, Coeff};
+
+pub(super) mod modulus;
 
 mod trivial;
-
-lazy_static! {
-    /// The polynomial modulus used for the polynomial field, `X^[MAX_POLY_DEGREE] + 1`.
-    /// This means that `X^[MAX_POLY_DEGREE] = -1`.
-    ///
-    /// This is the canonical but un-reduced form of the modulus, because the reduced form is the zero polynomial.
-    pub static ref POLY_MODULUS: DenseOrSparsePolynomial<'static, Coeff> = {
-        let mut poly = Poly::zero();
-
-        // Since the leading coefficient is non-zero, this is in canonical form.
-        // Resize to the maximum size first, to avoid repeated reallocations.
-        poly[MAX_POLY_DEGREE] = Coeff::one();
-        poly[0] = Coeff::one();
-
-        // Check canonicity and degree.
-        assert_eq!(poly.degree(), MAX_POLY_DEGREE);
-
-        poly.into()
-    };
-}
 
 /// A modular polynomial with coefficients in [`Coeff`], and maximum degree [`MAX_POLY_DEGREE`].
 #[derive(
@@ -64,14 +39,9 @@ lazy_static! {
     Into,
     Neg,
     Add,
-    // TODO: manually implement a final reduce step
-    Mul,
-    // TODO: manually implement a final reduce step
-    MulAssign,
+    // We don't implement DivAssign and RemAssign, because they have hidden clones.
     Div,
-    DivAssign,
     Rem,
-    RemAssign,
 )]
 pub struct Poly(DensePolynomial<Coeff>);
 
@@ -162,7 +132,7 @@ impl Index<usize> for Poly {
     /// Using `self.coeffs.iter()` is one way to ensure the code only accesses real indexes.
     fn index(&self, index: usize) -> &Self::Output {
         self.coeffs.get(index).expect(
-            "accessed redundant zero coefficient: \
+            "accessed virtual leading zero coefficient: \
             improve performance by stopping at the highest non-zero coefficient",
         )
     }
@@ -188,76 +158,41 @@ impl IndexMut<usize> for Poly {
     }
 }
 
+// We don't implement operators for SparsePolynomial or DenseOrSparsePolynomial, they are rare and can use .into() to convert first.
 impl Mul for Poly {
     type Output = Self;
 
     /// Multiplies then reduces by [`POLY_MODULUS`](static@POLY_MODULUS).
     fn mul(self, rhs: Self) -> Self {
-        let mut res = Self(&self.0 * &rhs.0);
-        res.reduce_mod_poly();
-        res
+        mul_poly(&self, &rhs)
     }
 }
 
-/// The fastest available modular polynomial operation.
-pub use mod_poly_manual_mut as mod_poly;
+impl Mul<&Poly> for Poly {
+    type Output = Self;
 
-/// Reduces `dividend` to `dividend % [POLY_MODULUS]`.
-///
-/// This is the most efficient manual implementation.
-pub fn mod_poly_manual_mut(dividend: &mut Poly) {
-    let mut i = MAX_POLY_DEGREE;
-    while i < dividend.coeffs.len() {
-        let q = i / MAX_POLY_DEGREE;
-        let r = i % MAX_POLY_DEGREE;
-
-        // In the cyclotomic ring we have that XˆN = -1,
-        // therefore all elements from N to 2N-1 are negated.
-        //
-        // For performance reasons, we use <Vec as IndexMut>,
-        // because the loop condition limits `i` to valid indexes.
-        if q % 2 == 1 {
-            dividend.coeffs[r] = dividend.coeffs[r] - dividend.coeffs[i];
-        } else {
-            dividend.coeffs[r] = dividend.coeffs[r] + dividend.coeffs[i];
-        }
-        i += 1;
+    /// Multiplies then reduces by [`POLY_MODULUS`](static@POLY_MODULUS).
+    fn mul(self, rhs: &Self) -> Self {
+        mul_poly(&self, rhs)
     }
-
-    // The coefficients of MAX_POLY_DEGREE and higher have already been summed above.
-    dividend.coeffs.truncate(MAX_POLY_DEGREE);
-
-    // The coefficients could sum to zero, so make sure the polynomial is in the canonical form.
-    dividend.truncate_to_canonical_form();
 }
 
-/// Returns the remainder of `dividend % [POLY_MODULUS]`, as a polynomial.
-///
-/// This clones then uses the manual implementation.
-pub fn mod_poly_manual_ref(dividend: &Poly) -> Poly {
-    let mut dividend = dividend.clone();
-    mod_poly_manual_mut(&mut dividend);
-    dividend
+impl Mul<DensePolynomial<Coeff>> for Poly {
+    type Output = Self;
+
+    /// Multiplies then reduces by [`POLY_MODULUS`](static@POLY_MODULUS).
+    fn mul(self, rhs: DensePolynomial<Coeff>) -> Self {
+        mul_poly(&self, &Self(rhs))
+    }
 }
 
-/// Returns the remainder of `dividend % [POLY_MODULUS]`, as a polynomial.
-///
-/// This uses an [`ark-poly`] library implementation, which always creates a new polynomial.
-pub fn mod_poly_ark_ref(dividend: &Poly) -> Poly {
-    let dividend: DenseOrSparsePolynomial<'_, _> = dividend.into();
+// TODO: if we need this method, remove the clone() using unsafe code
+#[cfg(inefficient)]
+impl Mul<&DensePolynomial<Coeff>> for Poly {
+    type Output = Self;
 
-    // The DenseOrSparsePolynomial implementation ensures canonical form.
-    let (_quotient, remainder) = dividend
-        .divide_with_q_and_r(&*POLY_MODULUS)
-        .expect("POLY_MODULUS is not zero");
-
-    remainder.into()
-}
-
-/// Reduces `dividend` to `dividend % [POLY_MODULUS]`.
-///
-/// This uses an [`ark-poly`] library implementation, and entirely replaces the inner polynomial representation.
-pub fn mod_poly_ark_mut(dividend: &mut Poly) {
-    let remainder = mod_poly_ark_ref(dividend);
-    *dividend = remainder;
+    /// Multiplies then reduces by [`POLY_MODULUS`](static@POLY_MODULUS).
+    fn mul(self, rhs: &DensePolynomial<Coeff>) -> Self {
+        mul_poly(&self, &Self(rhs.clone()))
+    }
 }
